# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Harry Potter API
  version: 1.0.0
  description: API for managing characters, magical creatures, and spells in the Harry Potter universe.
servers:
  - url: http://localhost:3000
    description: Development server
paths:
...
  /characters/{id}:
    get:
      summary: Get a character by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Character details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Character'
        '404':
          description: Character not found
    put:
      summary: Update a character
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewCharacter'
      responses:
        '200':
          description: Character updated
        '400':
          description: Invalid input
        '404':
          description: Character not found
    delete:
      summary: Delete a character
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Character deleted
        '404':
          description: Character not found
    patch:
      summary: Partially update a character
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CharacterPatch'
      responses:
        '200':
          description: Character partially updated
        '400':
          description: Invalid input
        '404':
          description: Character not found
...
components:
  schemas:
    Character:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        fullName:
          type: string
        house:
          type: string
        bloodStatus:
          type: string
          enum: [PURE_BLOOD, HALF_BLOOD, MUGGLE_BORN, UNKNOWN]
        wand:
          $ref: '#/components/schemas/Wand'
        patronus:
          type: string
        isDeathEater:
          type: boolean
        isOrderMember:
          type: boolean
        role:
          type: string
          enum: [STUDENT, TEACHER, AUROR, MINISTER_OF_MAGIC, DEATH_EATER, ORDER_MEMBER, OTHER]
        birthDate:
          type: string
          format: date
        deathDate:
          type: string
          format: date
        skills:
          type: array
          items:
            $ref: '#/components/schemas/Skill'
        magicalCreatures:
          type: array
          items:
            $ref: '#/components/schemas/CharacterCreatureRelationship'
        friends:
          type: array
          items:
            $ref: '#/components/schemas/Relationship'
        enemies:
          type: array
          items:
            $ref: '#/components/schemas/Relationship'
        notableEvents:
          type: array
          items:
            $ref: '#/components/schemas/Event'
        favoriteSpells:
          type: array
          items:
            type: string
...
    Wand:
      type: object
      properties:
        wood:
          type: string
        core:
          type: string
        length:
          type: number
        flexibility:
          type: string
    Skill:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        proficiency:
          type: string
          enum: [NOVICE, INTERMEDIATE, ADVANCED, MASTER]
    CharacterCreatureRelationship:
      type: object
      properties:
        creature:
          type: object
          properties:
            id:
              type: string
            type:
              type: string
        relationshipType:
          type: string
        details:
          type: string
    Relationship:
      type: object
      properties:
        character:
          $ref: '#/components/schemas/CharacterReference'
        relationshipType:
          type: string
    CharacterReference:
      type: object
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        fullName:
          type: string
        link:
          type: string
    Event:
      type: object
      properties:
        name:
          type: string
        date:
          type: string
          format: date
        description:
          type: string
...
    NewCharacter:
      type: object
      required:
        - firstName
        - lastName
      properties:
        firstName:
          type: string
        lastName:
          type: string
        house:
          type: string
        bloodStatus:
          type: string
        role:
          type: string
...
    CharacterPatch:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        house:
          type: string
        bloodStatus:
          type: string
        role:
          type: string
...
security:
  - apiKeyAuth: []
components:
  securitySchemes:
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key